fixes act error in console
